+++ rust/operator-binary/src/controller.rs
@@ -190,8 +190,10 @@ impl ReconcilerError for Error {
 pub async fn reconcile_hello(hello: Arc<HelloCluster>, ctx: Arc<Ctx>) -> Result<Action> {
     tracing::info!("Starting reconcile");
     let client = &ctx.client;
-    let resolved_product_image: ResolvedProductImage =
-        hello.spec.image.resolve(DOCKER_IMAGE_BASE_NAME);
+    let resolved_product_image: ResolvedProductImage = hello
+        .spec
+        .image
+        .resolve(DOCKER_IMAGE_BASE_NAME, crate::built_info::CARGO_PKG_VERSION);
     let hello_role = HelloRole::Server;
 
     let validated_config = validate_all_roles_and_groups_config(