+++ CHANGELOG.md
@@ -11,6 +11,11 @@ All notable changes to this project will be documented in this file.
 - Operator-rs: `0.44.0` ([#18]).
 - Increase the size limit of the log volumes ([#18])
 
+### Changed
+
+- Default stackableVersion to operator version ([#21]).
+
 [#14]: https://github.com/stackabletech/hello-world-operator/pull/14
 [#15]: https://github.com/stackabletech/hello-world-operator/pull/15
 [#18]: https://github.com/stackabletech/hello-world-operator/pull/18
+[#21]: https://github.com/stackabletech/hello-world-operator/pull/21